Renamed packages and removed some more stuff. Still breaks
