+++ SPR-9498/src/main/java/org/springframework/issues/DefaultFoo.java
@@ -11,23 +11,23 @@
  * subcomponent's license, as noted in the LICENSE file.
  */
 
-package org.springframework.security.oauth.examples.tonr.mvc;
+package org.springframework.issues;
 
-import java.util.Arrays;
-import java.util.Map;
-
-import org.springframework.util.LinkedMultiValueMap;
 
 /**
  * @author Dave Syer
- * 
+ *
  */
-public class DefaultBar extends LinkedMultiValueMap<String, String> implements Bar {
+public class DefaultFoo implements Foo {
 
-	public DefaultBar(Map<String, String[]> parameters) {
-		for (String key : parameters.keySet()) {
-			super.put(key, Arrays.asList(parameters.get(key)));
-		}
+	private Bar accessTokenRequest;
+
+	public DefaultFoo(Bar accessTokenRequest) {
+		this.accessTokenRequest = accessTokenRequest;
+	}
+	
+	public Bar getAccessTokenRequest() {
+		return accessTokenRequest;
 	}
 
 }
