+++ SPR-9498/src/main/java/org/springframework/issues/HomeController.java
@@ -1,4 +1,4 @@
-package org.springframework.security.oauth.examples.tonr.mvc;
+package org.springframework.issues;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
