+++ SPR-9498/src/main/webapp/WEB-INF/spring-servlet.xml
@@ -1,17 +1,15 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns:mvc="http://www.springframework.org/schema/mvc" xmlns:sec="http://www.springframework.org/schema/security"
-	xmlns:oauth="http://www.springframework.org/schema/security/oauth2"
-	xmlns:context="http://www.springframework.org/schema/context" xmlns:aop="http://www.springframework.org/schema/aop"
+	xmlns:mvc="http://www.springframework.org/schema/mvc" xmlns:context="http://www.springframework.org/schema/context"
+	xmlns:aop="http://www.springframework.org/schema/aop"
 	xsi:schemaLocation="http://www.springframework.org/schema/aop http://www.springframework.org/schema/aop/spring-aop-3.1.xsd
-		http://www.springframework.org/schema/security/oauth2 http://www.springframework.org/schema/security/spring-security-oauth2-1.0.xsd
 		http://www.springframework.org/schema/mvc http://www.springframework.org/schema/mvc/spring-mvc-3.1.xsd
 		http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.1.xsd
 		http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.1.xsd">
 
-	<bean id="conversionService" class="org.springframework.core.convert.support.DefaultConversionService"/>
+	<bean id="conversionService" class="org.springframework.core.convert.support.DefaultConversionService" />
 
-	<mvc:annotation-driven/>
+	<mvc:annotation-driven />
 
 	<!--Basic application beans. -->
 	<bean id="viewResolver" class="org.springframework.web.servlet.view.InternalResourceViewResolver">
@@ -19,18 +17,18 @@
 		<property name="suffix" value=".jsp" />
 	</bean>
 
-	<bean class="org.springframework.security.oauth.examples.tonr.mvc.HomeController">
-		<property name="clientContext" ref="foo"/>
+	<bean class="org.springframework.issues.HomeController">
+		<property name="clientContext" ref="foo" />
 	</bean>
 
-	<bean id="foo" class="org.springframework.security.oauth.examples.tonr.mvc.DefaultFoo" scope="request">
-		<aop:scoped-proxy proxy-target-class="false"/>
-		<constructor-arg ref="tokenRequest"/>
+	<bean id="foo" class="org.springframework.issues.DefaultFoo" scope="request">
+		<aop:scoped-proxy proxy-target-class="false" />
+		<constructor-arg ref="tokenRequest" />
 	</bean>
 
-	<bean id="tokenRequest" class="org.springframework.security.oauth.examples.tonr.mvc.DefaultBar" scope="request">
-		<aop:scoped-proxy proxy-target-class="false"/>
-		<constructor-arg value="#{request.parameterMap}"/>
+	<bean id="tokenRequest" class="org.springframework.issues.DefaultBar" scope="request">
+		<aop:scoped-proxy proxy-target-class="false" />
+		<constructor-arg value="#{request.parameterMap}" />
 	</bean>
 
 </beans>
